Conflicts:
	src/Starcounter.Apps.JsonPatch/BuiltInRestHandlers/SqlRestHandler.cs
	src/Starcounter.Bootstrap/Control.cs
	src/Starcounter.Internal/Http/Response.cs
	src/scnetworkgateway/OurSources/gateway.cpp

+++ src/Starcounter/Query/Optimization/_Interfaces.cs
@@ -19,7 +19,7 @@ internal interface IOptimizationNode
 
     void InstantiateNodesByExtentNumber(ExtentNode[] nodesByExtentNumber);
 
-    Int32 EstimateCost();
+    Double EstimateCost();
 
     List<IOptimizationNode> CreateAllPermutations();
